requirements.txt
httpx[http2]~=0.27
psycopg[binary,pool]~=3.1
python-dotenv~=1.0
